use chatgpt as the codeactagent
