--- evaluation/EDA/run_infer.sh
@@ -1,5 +1,0 @@
-python ./evaluation/EDA/run_infer.py \
---dataset things \
---data-split test \
---max-iterations 20 \
---eval-n-limit 1
